i think we should kill -Werror - it's brittle
git-svn-id: http://svn.aircrack-ng.org/trunk@1795 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/besside-ng.c
@@ -789,7 +789,7 @@ __no_resolve:
 		"Content-Length: %d\r\n\r\n",
 		_conf.cf_wpa_server,
 		boundary,
-		strlen(h1) + strlen(form) + tot);
+		(int) (strlen(h1) + strlen(form) + tot));
 
 	if (write(s, buf, strlen(buf)) != (int) strlen(buf))
 		goto __fail;
